requirements.txt
requests==2.19.1
google-auth==1.5.1
python-dateutil==2.7.3
pyyaml==4.2b4
oauthlib==3.0.0
kubernetes==8.0.1
click==6.7
click_log==0.3.2
docker==3.7.0
nsenter==0.2
attrs==19.1.0
pluggy==0.11.0
pytest==4.5.0
urllib3<1.24,>=1.21.1
setuptools>38.3